renamed userBookingHistory file to a meaningful name
